+++ src/component/atoms/Text.js
